+++ modules/v1_20_R1/src/main/java/net/countercraft/movecraft/support/v1_20_R1/IAsyncChunk.java
@@ -1,4 +1,4 @@
-package net.countercraft.movecraft.support.v1_19_R3;
+package net.countercraft.movecraft.support.v1_20_R1;
 
 import com.google.common.cache.CacheBuilder;
 import com.google.common.cache.CacheLoader;
